refactor(aiops): split AIops (506→108) em hook + componentes
Extrai verbatim de src/pages/AIops.tsx:
- useAiOps: 4 hooks de dados (decisões/perfis/rodar agente/atualizar status)
  + filtros + particionamento prioridades/oportunidades/riscos
- AiOpsHeader, StatsCards, DecisionList (reutilizado nas 3 abas, idênticas
  exceto ícone/mensagem do empty state), DecisionCard, EvidenceItem,
  types.ts, config.ts (actionIcons/actionLabels/confidenceBadge)
- +13 testes (config + presentacionais)

Comportamento preservado 1:1. tsc baseline=0, eslint=0.
